init work - handscomp is unable to be pulled
