+++ Content.Server/VoiceTrigger/StorageVoiceControlComponent.cs
@@ -4,4 +4,4 @@ namespace Content.Server.VoiceTrigger;
 /// entities with this component, ItemSlots, and TriggerOnVoiceComponent will eject the name of the item slot spoken after the TriggerOnVoiceComponent has been activated
 /// </summary>
 [RegisterComponent]
-public sealed partial class ItemSlotVoiceControlComponent : Component;
+public sealed partial class StorageVoiceControlComponent : Component;
